feat(#1598): extract DcsEachWithoutTransitive class
